dopuna

+++ FrontendApp/GuiRadnici/GuiRadnici/racun.cs
@@ -10,8 +10,8 @@ namespace GuiRadnici
     {
         public int idracuna { get; set; }
         public DateTime datum { get; set; }
-        public double ukupno { get; set; }
-        //public string radnik_jmb { get; set; }
+        public decimal ukupno { get; set; }
+       
         public radnik radnik { get; set; }
 
     }
